Version 0.8.3
- Added Default Configuration to Admin Menu
- Added the option to have no notification
- Added Home System
- Added More Flags
- Fixed null pointer exception when running certain events and a player is not a member of a settlement
- Fixed Accept command adding trusted members
- Fixed Auto claim logic
- Fixed Claim command logic
- Fixed Claims inventory logic
- Fixed Some flags from not working
- Fixed max claims not updating correctly
- Fixed permission checking for cetain commands
- Improved Notification Selection in menu
- Improved Modules
- Improved Commands
- Lots more bug fixes

+++ src/main/java/com/huskydreaming/settlements/inventories/providers/FlagsInventory.java
@@ -4,10 +4,11 @@ import com.huskydreaming.huskycore.HuskyPlugin;
 import com.huskydreaming.huskycore.inventories.InventoryItem;
 import com.huskydreaming.huskycore.inventories.InventoryPageProvider;
 import com.huskydreaming.settlements.enumeration.Flag;
+import com.huskydreaming.settlements.enumeration.RolePermission;
+import com.huskydreaming.settlements.services.interfaces.*;
 import com.huskydreaming.settlements.storage.persistence.Member;
-import com.huskydreaming.settlements.services.interfaces.FlagService;
-import com.huskydreaming.settlements.services.interfaces.InventoryService;
-import com.huskydreaming.settlements.services.interfaces.MemberService;
+import com.huskydreaming.settlements.storage.persistence.Role;
+import com.huskydreaming.settlements.storage.persistence.Settlement;
 import fr.minuskube.inv.content.InventoryContents;
 import org.bukkit.entity.Player;
 import org.bukkit.event.inventory.InventoryClickEvent;
@@ -15,18 +16,22 @@ import org.bukkit.inventory.ItemStack;
 
 public class FlagsInventory extends InventoryPageProvider<Flag> {
 
-    private final FlagService flagService;
     private final HuskyPlugin plugin;
-    private final MemberService memberService;
+    private final FlagService flagService;
     private final InventoryService inventoryService;
+    private final MemberService memberService;
+    private final RoleService roleService;
+    private final SettlementService settlementService;
 
     public FlagsInventory(HuskyPlugin plugin, int rows, Flag[] flags) {
         super(rows, flags);
         this.plugin = plugin;
 
-        this.flagService = plugin.provide(FlagService.class);
-        this.inventoryService = plugin.provide(InventoryService.class);
-        this.memberService = plugin.provide(MemberService.class);
+        flagService = plugin.provide(FlagService.class);
+        inventoryService = plugin.provide(InventoryService.class);
+        memberService = plugin.provide(MemberService.class);
+        roleService = plugin.provide(RoleService.class);
+        settlementService = plugin.provide(SettlementService.class);
     }
 
     @Override
@@ -38,15 +43,34 @@ public class FlagsInventory extends InventoryPageProvider<Flag> {
 
     @Override
     public ItemStack construct(Player player, int i, Flag flag) {
+        if (!memberService.hasSettlement(player)) {
+            player.closeInventory();
+            return null;
+        }
+
         Member member = memberService.getCitizen(player);
+        Role role = roleService.getRole(member);
+        Settlement settlement = settlementService.getSettlement(member.getSettlement());
+
         boolean enabled = flagService.hasFlag(member.getSettlement(), flag);
+        if (!(role.hasPermission(RolePermission.EDIT_FLAGS) || settlement.isOwner(player))) {
+            return InventoryItem.of(enabled, false, flag.toString(), flag.getDescription());
+        }
+
         return InventoryItem.of(enabled, flag.toString(), flag.getDescription());
     }
 
     @Override
     public void run(InventoryClickEvent event, Flag flag, InventoryContents contents) {
         if (event.getWhoClicked() instanceof Player player) {
+            if (!memberService.hasSettlement(player)) return;
+
             Member member = memberService.getCitizen(player);
+            Role role = roleService.getRole(member);
+            Settlement settlement = settlementService.getSettlement(member.getSettlement());
+
+            if (!(role.hasPermission(RolePermission.EDIT_FLAGS) || settlement.isOwner(player))) return;
+
             boolean enabled = flagService.hasFlag(member.getSettlement(), flag);
             if (enabled) {
                 flagService.removeFlag(member.getSettlement(), flag);
